Reworked the shockwaveprojectile (sort of)
The initial projectile now detonates on walls and other tank instead of
just be destroyed. This should make the weapon a little bit more useful
(and fun). Also, the initial projectile no longer deals any damage. The
damage of the shockwave itself might have to be tweaked since it's a
little overpowered at the moment. A solution might have to be
implemented where there's a maximum damage that the shockwave can do to
a single tank.